Add support for live preview in capture mode

+++ LongoMatch.GUI.Multimedia/Gui/PlayerBin.cs
@@ -105,9 +105,9 @@ namespace LongoMatch.Gui
 			}
 		}
 
-		public long CurrentTime {
+		public int CurrentTime {
 			get {
-				return player.CurrentTime;
+				return (int) player.CurrentTime;
 			}
 		}
